requirements.txt
pip~=24.2
attrs~=24.2.0
requests~=2.32.3
python-dotenv~=1.0.1